refactor data logging and start using sparse dataframes for tick logs

+++ src/stellarpunk/core.py
@@ -88,7 +88,7 @@ class ProductionChain:
                 self.value_estimate[product_id], self.volume_estimate[product_id],
                 self.price_vema_alpha, price, volume)
 
-    def inputs_of(self, product_id:int) -> npt.NDArray[np.float64]:
+    def inputs_of(self, product_id:int) -> npt.NDArray[np.int64]:
         return np.nonzero(self.adj_matrix[:,product_id])[0]
 
     def first_product_ids(self) -> npt.NDArray[np.int64]:
